change buffer lengths of sysctl(2) to be size_t's
SCCS-vsn: sys/kern/uipc_domain.c 7.14
SCCS-vsn: sys/vm/vm_meter.c 7.20

+++ usr/src/sys/kern/uipc_domain.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)uipc_domain.c	7.13 (Berkeley) %G%
+ *	@(#)uipc_domain.c	7.14 (Berkeley) %G%
  */
 
 #include <sys/param.h>
@@ -113,9 +113,9 @@ net_sysctl(name, namelen, oldp, oldlenp, newp, newlen, p)
 	int *name;
 	u_int namelen;
 	void *oldp;
-	u_int *oldlenp;
+	size_t *oldlenp;
 	void *newp;
-	u_int newlen;
+	size_t newlen;
 	struct proc *p;
 {
 	register struct domain *dp;
